requirements.txt
Django==2.2.1
graphene-django==2.2.0
==0.2.1
django-graphql-jwt